tutorials/deploy-layer-models/gradio/requirements.txt
layer-sdk
torch==1.11.0
sentencepiece==0.1.96